# openapi.yaml
openapi: 3.0.3
info:
  title: LeadRecovery API
  description: >
    Multi-tenant lead recovery and reactivation API. Every route except
    /health, /ready, and the provider webhooks requires
    `Authorization: Bearer <tenant api key>`; admin routes require
    `Authorization: Bearer <ADMIN_API_KEY>` instead. See README.md for the
    full narrative reference — this file is the machine-readable version of
    the same routes.
  version: "0.2.0"
servers:
  - url: http://localhost:3000
    description: Local demo server (npm run dev)
...
security: []
paths:
...
  /leads:
    get:
      tags: [leads]
      summary: List the tenant's leads
      security: [{ tenantAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: needsAttention
          in: query
          description: When "true", restricts the list to leads currently flagged in the needs-attention inbox (see POST /leads/{id}/reply and /leads/{id}/mark-handled).
          schema: { type: string, enum: ["true"] }
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count: { schema: { type: integer } }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Lead" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Suspended" }
  /leads/plan:
    get:
      tags: [leads]
      summary: Dry run — scored + composed plans, nothing sent
      security: [{ tenantAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count: { schema: { type: integer } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkflowResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Suspended" }
  /leads/{id}/messages:
    get:
      tags: [leads]
      summary: Conversation history for one lead
      security: [{ tenantAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Message" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Suspended" }
  /leads/{id}:
    patch:
      tags: [leads]
      summary: Update one lead's editable fields
      description: >
        Never accepts `status` (compliance-sensitive; only ever set by the
        classify/workflow logic) — rejected with 400. Clearing
        `appointmentAt` also clears `appointmentReminderSentAt`, so
        rescheduling gets a fresh 24-hours-before reminder instead of being
        silently blocked by the old appointment's sent-flag.
      security: [{ tenantAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadPatchBody" }
      responses:
        "200":
          description: The updated lead
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lead" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Suspended" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [leads]
      summary: Permanently delete one lead (a right-to-erasure request)
      description: >
        Cascades to the lead's message history in Postgres. No undo.
        Distinct from the automatic data-retention purge (see
        COMPLIANCE.md), which never touches do_not_contact/opted_out leads
        — this endpoint deletes regardless of status, since it's an
        explicit request for one specific person's data.
      security: [{ tenantAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Suspended" }
        "404": { $ref: "#/components/responses/NotFound" }
  /leads/{id}/convert:
    post:
      tags: [leads]
      summary: Mark a lead as a converted/won customer
      description: >
        Sets `status: "converted"` and `convertedAt`. Optionally records this
        specific lead's own `marketingOptIn` for later win-back messaging —
        see COMPLIANCE.md "Win-back messaging for past customers". Converting
        a lead does not by itself opt it into anything; `marketingOptIn`
        must be explicitly true (here or via a later `PATCH /leads/{id}`).
      security: [{ tenantAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadConvertBody" }
      responses:
        "200":
          description: The updated lead
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lead" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Suspended" }
        "404": { $ref: "#/components/responses/NotFound" }
  /leads/{id}/reply:
    post:
      tags: [leads]
      summary: Send a real reply on an operator's behalf, resolving a needs-attention item
      description: >
        Sends through the lead's usual channel (selectChannel's own choice)
        unless `channel` overrides it. Logged as kind "manual_reply". Clears
        needsAttentionAt/needsAttentionReason/attentionAlertedAt on success.
        A "chat" reply is recorded but cannot be pushed live into an
        already-open chat-widget tab (no WebSocket/SSE infra) — it surfaces
        next time that visitor's widget polls.
      security: [{ tenantAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadReplyBody" }
      responses:
        "200":
          description: The updated lead
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lead" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Suspended" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { description: "No usable channel for this lead, or carrier approval is still required" }
        "502": { description: "The provider send itself failed" }
  /leads/{id}/mark-handled:
    post:
      tags: [leads]
      summary: Clear a needs-attention flag without sending anything
      description: >
        For when an operator resolved the lead some other way (a phone call,
        etc.) — just clears needsAttentionAt/needsAttentionReason/attentionAlertedAt.
      security: [{ tenantAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The updated lead
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lead" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Suspended" }
        "404": { $ref: "#/components/responses/NotFound" }
  /leads/import:
    post:
      tags: [leads]
      summary: Bulk-create leads from a CSV file
      description: >
        Same columns/validation as `npm run import-leads` (see
        README.md/ONBOARDING.md): name, phone, email, source, createdAt,
        requestedService, previousQuote, appointmentAt, notes — at least
        phone or email is required per row, and other rows are skipped
        (counted, not rejected). A row with a valid appointmentAt is
        automatically marked appointmentStatus "booked". Skips (rather than
        merging) a row whose phone/email already matches an existing lead
        for this tenant, or an earlier row in the same file — counted
        separately as `duplicates`.
      security: [{ tenantAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [csv]
              properties:
                csv: { type: string, description: "Raw CSV file contents, max 5MB." }
      responses:
        "201":
          description: Import result
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported: { type: integer }
                  skipped: { type: integer, description: "Rows with neither phone nor email." }
                  duplicates: { type: integer, description: "Rows matching an existing lead, or an earlier row in this same file." }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Suspended" }
  /leads/export:
    get:
      tags: [leads]
      summary: Full-fidelity export of every lead field
      description: >
        A one-shot download (never paginated) — for a client's own records
        or a data right-of-access request. CSV by default; pass
        ?format=json for a plain JSON array instead.
      security: [{ tenantAuth: [] }]
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [csv, json], default: csv }
      responses:
        "200":
          description: OK
          content:
            text/csv:
              schema: { type: string }
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Lead" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Suspended" }
...
components:
  securitySchemes:
    tenantAuth:
      type: http
      scheme: bearer
      description: A tenant's own API key.
    adminAuth:
      type: http
      scheme: bearer
      description: ADMIN_API_KEY.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1 }
    Offset:
      name: offset
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
  responses:
    Unauthorized:
      description: Missing or invalid Authorization header
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Suspended:
      description: Tenant is suspended
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Lead:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        source:
          type: string
          enum: [crm, website_form, missed_call, booking_software, email, sms, whatsapp, chat, spreadsheet, customer_database, other]
        createdAt: { type: string, format: date-time }
        status:
          type: string
          enum:
            [
              new,
              contacted_no_response,
              responded,
              booked,
              converted,
              do_not_contact,
              unqualified,
              fraudulent,
              active_conversation,
              opted_out,
            ]
        requestedService: { type: string }
        previousQuote: { type: string }
        appointmentStatus:
          type: string
          enum: [none, requested, abandoned, booked]
        appointmentAt:
          type: string
          format: date-time
          description: "When the booked appointment is scheduled for. Only meaningful when appointmentStatus is \"booked\"; drives the 24-hours-before reminder."
        appointmentReminderSentAt:
          type: string
          format: date-time
          description: "When the appointment reminder was actually sent — set automatically, guards against sending it twice."
        preferredChannel:
          type: string
          enum: [sms, whatsapp, email]
        notes: { type: string }
        convertedAt:
          type: string
          format: date-time
          description: When this lead's status was set to "converted" — see POST /leads/{id}/convert.
        marketingOptIn:
          type: boolean
          description: This specific lead's own opt-in to win-back messaging after converting — see COMPLIANCE.md "Win-back messaging for past customers".
        lastWinBackAt:
          type: string
          format: date-time
          description: When the last win-back check-in was actually sent.
        chatTokenHash:
          type: string
          description: Hash only (never the raw token) of this lead's chat-widget session token, if it was created that way. See POST /public/chat/{tenantId}/start.
        needsAttentionAt:
          type: string
          format: date-time
          description: Set when this lead needs a human's attention (an "interested" reply, or the auto-reply chatbot escalating) — drives the dashboard's needs-attention inbox. Never overwritten by a later inbound message while already set, so it reflects true wait time. Cleared by POST /leads/{id}/reply or /leads/{id}/mark-handled.
        needsAttentionReason:
          type: string
          enum: [interested, needs_human_reply]
          description: Why needsAttentionAt is set.
        attentionAlertedAt:
          type: string
          format: date-time
          description: Set once an SLA operator-alert has fired for this specific unresolved needsAttentionAt, so it only fires once. Cleared along with needsAttentionAt.
    LeadPatchBody:
      type: object
      description: Fields accepted by PATCH /leads/{id}. Never accepts status.
      properties:
        name: { type: string }
        requestedService: { type: string }
        previousQuote: { type: string }
        notes: { type: string }
        preferredChannel:
          type: string
          enum: [sms, whatsapp, email]
        appointmentStatus:
          type: string
          enum: [none, requested, abandoned, booked]
        appointmentAt:
          type: string
          format: date-time
          nullable: true
          description: Set null to clear it (also clears appointmentReminderSentAt).
        marketingOptIn:
          type: boolean
          description: This lead's own opt-in to later win-back messaging.
    LeadConvertBody:
      type: object
      description: Body for POST /leads/{id}/convert.
      properties:
        marketingOptIn:
          type: boolean
          description: Whether this lead personally agreed to occasional win-back messages after becoming a customer. Omitted/false means not opted in.
    LeadReplyBody:
      type: object
      description: Body for POST /leads/{id}/reply.
      required: [message]
      properties:
        message: { type: string, description: The reply text to send. }
        channel:
          type: string
          enum: [sms, whatsapp, email, chat]
          description: Overrides the automatically-selected channel (selectChannel's own choice, same as an automated send).
    Message:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        leadId: { type: string }
        channel: { type: string, enum: [sms, whatsapp, email] }
        direction: { type: string, enum: [outbound, inbound] }
        body: { type: string }
        at: { type: string, format: date-time }
        classification: { type: string, enum: [stop, interested, not_interested, question, unknown] }
        deliveryStatus: { type: string }
        kind: { type: string, enum: [campaign, auto_reply, closer, appointment_reminder, win_back, manual_reply] }
    RecoveryPlan:
      type: object
      properties:
        lead: { $ref: "#/components/schemas/Lead" }
        priority: { type: string, enum: [HIGH, MEDIUM, LOW] }
        priorityReasons: { type: array, items: { type: string } }
        reason:
          type: object
          properties:
            text: { type: string }
            grounded: { type: boolean }
        message:
          type: object
          properties:
            channel: { type: string, enum: [sms, whatsapp, email] }
            body: { type: string }
    WorkflowResult:
      type: object
      properties:
        plans: { type: array, items: { $ref: "#/components/schemas/RecoveryPlan" } }
        skipped:
          type: array
          items:
            type: object
            properties:
              lead: { $ref: "#/components/schemas/Lead" }
              reason: { type: string }
...